requirements.txt
django==1.11.1
djangorestframework==3.6.3
markdown==2.6.8
django-filter==1.0.2